fix(answer-retrospective) create fetchSummaryData hook
