+++ src/components/Routine/hooks/useAnswerSummaryPagination/use-answer-summary-pagination.tsx
@@ -1,36 +1,16 @@
-import { useCallback, useContext } from 'react'
-import { useRecoilValue, useSetRecoilState } from 'recoil'
+import { useRecoilValue } from 'recoil'
 
-import { ServicesContext } from 'src/components/Base/ServicesProvider/services-provider'
 import { Team } from 'src/components/Team/types'
 import { User } from 'src/components/User/types'
 import { answerSummaryAtom } from 'src/state/recoil/routine/answer-summary'
-import { answerSummaryPaginationAtom } from 'src/state/recoil/routine/cursor-answer-summary-pagination'
 import { filteredUsersCompany, selectUserFromCompany } from 'src/state/recoil/team/users-company'
 import meAtom from 'src/state/recoil/user/me'
 
-import useAnswerSummaryFormatter from '../../RetrospectiveTab/Answers/utils/answer-summary-formatter'
-import { AnswerSummary, formatUUIDArray } from '../../RetrospectiveTab/retrospective-tab-content'
-
-type fetchAnswersProperties = {
-  teamUsersIds: string[]
-  after: Date
-  before: Date
-}
-
 type AnswerSummaryPaginationProperties = {
   limitedTeamUsers: User[]
-  fetchAnswers: ({
-    after,
-    before,
-    teamUsersIds,
-  }: fetchAnswersProperties) => Promise<AnswerSummary[]>
 }
 
 const useAnswerSummaryPagination = (teamId: Team['id']): AnswerSummaryPaginationProperties => {
-  const setAnswerSummaryPaginationData = useSetRecoilState(answerSummaryPaginationAtom)
-  const { servicesPromise } = useContext(ServicesContext)
-  const { formattedAnswerSummary } = useAnswerSummaryFormatter()
   const teamUsers = useRecoilValue(filteredUsersCompany(teamId))
   const userID = useRecoilValue(meAtom)
   const me = useRecoilValue(selectUserFromCompany(userID))
@@ -50,39 +30,7 @@ const useAnswerSummaryPagination = (teamId: Team['id']): AnswerSummaryPagination
 
   const limitedTeamUsers = filteredTeamUsers.slice(0, REQUEST_LIMIT)
 
-  const fetchAnswers = useCallback(
-    async ({ after, before, teamUsersIds }: fetchAnswersProperties): Promise<AnswerSummary[]> => {
-      const { routines } = await servicesPromise
-
-      const parsetToQueryTeamUsersIDS = encodeURIComponent(formatUUIDArray(teamUsersIds))
-
-      setAnswerSummaryPaginationData({
-        lastLoadedUserId: teamUsersIds[teamUsersIds.length - 1],
-        teamId,
-      })
-      const { data: answersSummaryData } = await routines.get<AnswerSummary[]>(
-        `/answers/summary/${teamId}`,
-        {
-          params: {
-            before,
-            after,
-            includeSubteams: false,
-            teamUsersIds: parsetToQueryTeamUsersIDS,
-          },
-        },
-      )
-
-      const formattedData = formattedAnswerSummary({
-        requestedUsersIDs: teamUsersIds,
-        answerSummary: answersSummaryData,
-      })
-
-      return formattedData
-    },
-    [formattedAnswerSummary, servicesPromise, setAnswerSummaryPaginationData, teamId],
-  )
-
-  return { limitedTeamUsers, fetchAnswers }
+  return { limitedTeamUsers }
 }
 
 export default useAnswerSummaryPagination
