+++ src/components/Routine/hooks/useFetchSummaryData/index.ts
@@ -1,0 +1,1 @@
+export { default as useFetchSummaryData } from './use-fetch-summary-data'
